fungsi baru

+++ database/migrations/2023_09_29_044744_create_mapels_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('judul');
             $table->string('image');
-            $table->decimal('harga');
+            $table->integer('harga');
             $table->timestamps();
         });
     }
